Running asyncio apps in click and recursion limit of c
Would you know it, running asynchronous programs using click is as easy
as it can get. There are lots of beautiful patterns to do this, my
favorite on is using a an executor thread, a server and using run in
asyncio and uvloop to run coroutines to get back the results as if they
were synchronous. The first one involves starting a thread and starting
an asyncio event loop in it where you dispatch tasks to it via functions
like `run_coroutine_threadsafe` and `to_thread` which heavily utilize
concurrent.futures library and you get the benefits of mixing async and
sync seamlessly. The second one involves running your asyncronous and
synchronous applications in different processes and have them
communicate via sockets and network connections, the unix-socket is best
for this, where you can serialize objects using pickle and send them to
the server for processing and wait for the server to communicate back
the result, amazingly simple, this kind of architecture involves using
task ids such that when the results are communicated back, they are
accompanied by the id so that you can associate it with its
corresponding task and react as expected. The third is self explanatory,
the run function in uvloop and asyncio return the result of the main
coroutine passed to it, the biggest issue with it is does each call to
run create new EventLoops, because this can get heavy real fast and
don't get me started with the devilish bugs accompanied by them.
Otherwise, I mentioned C/C++ recursion limit, more of the callstack
limit, from my simple test with the 4 compilers I have,
clang/clang++/gcc/g++ they all yield different results even for the
executables, they produce different results on every run, and my
conclusion is that the limit is around 261700, pretty big and the
program is killed by SEGV interrupt when it is exceeded. This means for
recursions that run to this amount are mostly errors in my programs and
hence I can rely on a few thousand recursions before I get to my result.
NIGHT.