Python metaclasses, C++ templates types and lookup
Today is on Sunday, and I thought I will make this day more about
understanding some core gory language details like c++ templates, python
metaclasses and lookups. I new classes were instances as are there
instaces but I never new/saw that literally. Classes behave like there
instances as far as my tests have shown me. What do I know? Classes that
inherit from `type` create other classes as their instances in which
those classes create the objects you and I mostly use. I also know that
for attribute lookup in instances starts at the instance dictionary
(where instance variables live), if not found here it is checked in its
class' dictionary (where methods and class variables live), if not
found, then the dictionary of the superclass is checked, if also not
found in here, its superclass is checked (the superclass of the
superclass of the original instance's class) all the way up until `type`
is reached which I believe raises the AttributeError. My understanding
is so miniscule since I cannot figure out where `__getattr__` is called
but I will investigate this later. What I have just defined is the
lookup of a instance that you and I use, this is the good ol' literal
string, an instance of the `str` class. But remember classes are also
instances and thus this applies too. I have noticed that classes can
have variables that there instances cannot access since they are stored
in the class' metaclass, and the instance lookup does not go that far to
look for a name. Hence, the superclass chain of normal instances are normal
classes, while the superclass chain of classes are metaclasses (classes
that inherit from `type`), classic. This is just mind blowing. I believe
true class methods are defined in the metaclass since only by
`class.method` can the method be accessed and also it returns a bound
method like a normal instance. This means `instance.method` will raise
an AttributeError. This means class variables are set in the metaclass
`__init__` method. MindBlown.

Instances look for methods in their class type, which means classes
look for their methods in their metaclass type, since metaclasses are
not in the search path of instances then functions defined in
metaclasses (class methods) can only be accessed via the class not its
instances but class variables are accessible from its instances since
they live in the class' dictionary which is in the instances lookup
path. Hence, metaclasses functions are true class methods. The
`classmethod` decorator is simply a workaround using descriptors instead
of defining a whole metaclass just to give a class a method, but
remember this makes the classmethod live on the class' dictionary hence
accessible to instances therefore not true class methods. The truthiness
stated here is about who owns what and has the rights by definition of
use. Class methods should be accessible to classes not their instances
but again all instances know there classes therefore on calling a
`classmethod` it can substitute its class for self instead of itself thus
a class method which I also think is how `classmethod` works.

Otherwise, I revisited some good ol' templates. C++ currently supports 4 types of
templates: variable, function, alias and class templates. I have played
around with them in their proper translation units, check it out. User
defined infix operators are possible in c++ as I have just done in the
templates sources, but unfortunately, due to their definition, it is
hard to use in variadic template folding, I will look into this some
other time. That's all folks, metaprogramming, a word so beautiful and
yet so complexly vague and full of abstractions. It's amazing that
almost all of the type_traits/concepts in c++ relies on true and false.
Will be back with more mind blowing facts. NIGHT.