Read some python c-implementation stuff
Today i explored some python c-implementation code. I started this some
time back and I had a hard time understanding anything in there but
today after spending the whole day looking at it and wrapping my head
around ieee754 spec i finally had a snap. I was able to compile
python3.13 and tested it and also found some DSL within python that is
used to generate code for the interpreter and interestingly, python is
used to create python and the techniques used to lex, parse and analyze
the DSL were well implemented, they were very similar to LL(k) and the
other powerful ones. Turns out, there are alot of code generators in the
python repo from generating makefiles, interpreter, configurations, and
creating executables that depend on the system in which they are run,
ie, they gather infomation about the system and use it to generate other
files. The level of metaprogramming going on in the python repo is
outstanding, the only thing we have are the core types the rest are
generated by tools like autoconf, pegen and the custom DSL and many more
that I've not recognized. I'll look into the techniques used in the DSL
to learn more about parsers in my other projects. Night.