Test Python Limit
Talk about flexibility, python is well designed. According to how local
variables are referenced in bytecode, I thought the most local variables
a functiopn can have in python is no more than 255 since it is stored as
a byte after the LOAD_FAST instruction but to my surprise, you can store
way more than you could ever need, possibly `2**32` since python uses
modifier instructions to build the location index which is a 32 bit
integer. Unfortunately, creating a function with `2**32` parameters is
easy but parsing it becomes a problem, it hangs my computer and
sometimes craches my terminal. Not only that, the python interpreter
stack can handle some very big number of local variables. Though there
are lots of instructions in there that I just can't get to understand
their purpose, example the PRECALL which is executed before every CALL
instruction. Before I go, it is amazing how python handles chained
assignments, first the value is computed and pushed to the stack, then a
COPY instruction is executed that peeks the value on the top, duplicates
it and pushes the duplicate on the stack then the
STORE_{LOCAL,GLOBAL,...} instruction pops a value from the stack and
stores it where necessary, it repeats this process until only one
assignment is left, which takes the original value hence the assignment
statement has a zero stack effect. And by the way, the walrus operator
is not so mystical either, like the assignment, the walrus compiles to a
STORE_{FAST,GLOBAL,...} instrction but the catch is that there is the
good ol' COPY which duplicates the value on the stack, then one is
popped and stored appropriately with the STORE_\* instruction, the
remaining original one is left as the result of the expression since all
expressions have a non-zero stack effect. Otherwise, Night.