How monitors work
Read on how monitors work, how colors and images are represented on the
screen which acts as the canvas. I implemented a simply naive canvas and
a basic shape, the rectangle, which can be used to create simple images
like a face and flag as shown by the `imagen.py` program along with an
empty canvas. More will be added, example, serializing an image to bytes
and deserializing back to a viewable image, all part of bit
manipulation.