Added move/copy semantics to variant
Along with emplace, we have made the variant much more efficiant by
enabling moving and copying the current alternative. Move/copy apply
only to the current alternative, not the variant, after, say a move, the
moved variant is left with an empty alternative while the constructed
move receives the resources of the current alternative, the type in the
receiver is also updated respectively. The emplace makes the whole
structure much more efficient by enebling in place construction in the
variant.