Better way of storing size as an integer
Using bit mask in every consecutive 10 bits, you can store the size of
something in bytes, kilobytes, megabytes, gigabytes, terabytes, and
petabytes. These correspond with the 64 bits of the uint64_t. 64 bits
can encode to a maximum of 1023 (bytes, KB, MB, GB, TB, and PB) and 7
EB. This kind of memory cannot really be found on any typical computer
but it is amazing how consecutive 10 bits can encode how much of the
different byte-bundles the size holds. to think my earlier
implementation `mpack.bsize` was effective.