Where's c++ realloc? well, std::vector!
So, insearch of c++'s realloc, I concluded it is std::vector, and to
confirm this, I implemented a working vector using c's realloc, now
std::realoc from `<cstdlib>`. It so happens to fit what std::vector does
and therefore std::vector might be of the same nature in implementation
if not with a few optimizations of its own like the infamous
std::vector<bool> specialization.